Refactor to components folder
