+++ app/home/components/contactBlock.tsx
@@ -1,0 +1,11 @@
+import IconBlock from "./iconBlock";
+
+export default function ContactBlock() {
+  return (
+    <div className="bg-red-600">
+      <IconBlock iconPath="" content="a" />
+      <IconBlock iconPath="" content="b" />
+      <IconBlock iconPath="" content="c" />
+    </div>
+  );
+}
