+++ app/home/components/textBlock.tsx
