--- app/home/contactBlock.tsx
@@ -1,8 +1,0 @@
-export default function ContactBlock() {
-  return (
-    <div>
-      <p>Wanna Talk?</p>
-      <div></div>
-    </div>
-  );
-}
